refactor(shorebird_cli): delete old patch command and mixins
