+++ packages/shorebird_cli/lib/src/commands/patch/android_patcher.dart
@@ -1,50 +1,44 @@
-import 'package:archive/archive_io.dart';
 import 'package:crypto/crypto.dart';
 import 'package:io/io.dart';
-import 'package:mason_logger/mason_logger.dart';
 import 'package:path/path.dart' as p;
 import 'package:shorebird_cli/src/archive_analysis/android_archive_differ.dart';
 import 'package:shorebird_cli/src/archive_analysis/archive_differ.dart';
 import 'package:shorebird_cli/src/artifact_builder.dart';
 import 'package:shorebird_cli/src/artifact_manager.dart';
 import 'package:shorebird_cli/src/code_push_client_wrapper.dart';
-import 'package:shorebird_cli/src/commands/patch_new/patch_new.dart';
+import 'package:shorebird_cli/src/commands/patch/patcher.dart';
+import 'package:shorebird_cli/src/doctor.dart';
 import 'package:shorebird_cli/src/logger.dart';
 import 'package:shorebird_cli/src/patch_diff_checker.dart';
 import 'package:shorebird_cli/src/platform.dart';
 import 'package:shorebird_cli/src/platform/platform.dart';
 import 'package:shorebird_cli/src/release_type.dart';
 import 'package:shorebird_cli/src/shorebird_android_artifacts.dart';
-import 'package:shorebird_cli/src/shorebird_env.dart';
 import 'package:shorebird_cli/src/shorebird_flutter.dart';
 import 'package:shorebird_cli/src/shorebird_validator.dart';
 import 'package:shorebird_cli/src/third_party/flutter_tools/lib/flutter_tools.dart';
 import 'package:shorebird_cli/src/version.dart';
 import 'package:shorebird_code_push_client/shorebird_code_push_client.dart';
 
-/// {@template aar_patcher}
-/// Functions to patch an AAR release.
+/// {@template android_patcher}
+/// Functions to create an Android patch.
 /// {@endtemplate}
-class AarPatcher extends Patcher {
-  /// {@macro aar_patcher}
-  AarPatcher({
+class AndroidPatcher extends Patcher {
+  /// {@macro android_patcher}
+  AndroidPatcher({
     required super.argResults,
     required super.flavor,
     required super.target,
   });
 
-  /// The build number of the aar (1.0). Forwarded to the --build-number
-  /// argument of the flutter build aar command.
-  String get buildNumber => argResults['build-number'] as String;
+  @override
+  ReleaseType get releaseType => ReleaseType.android;
+
+  @override
+  String get primaryReleaseArtifactArch => 'aab';
 
   @override
   ArchiveDiffer get archiveDiffer => AndroidArchiveDiffer();
-
-  @override
-  String get primaryReleaseArtifactArch => 'aar';
-
-  @override
-  ReleaseType get releaseType => ReleaseType.aar;
 
   @override
   Future<void> assertPreconditions() async {
@@ -52,37 +46,51 @@ class AarPatcher extends Patcher {
       await shorebirdValidator.validatePreconditions(
         checkUserIsAuthenticated: true,
         checkShorebirdInitialized: true,
+        validators: doctor.androidCommandValidators,
       );
     } on PreconditionFailedException catch (e) {
       exit(e.exitCode.code);
-    }
-
-    if (shorebirdEnv.androidPackageName == null) {
-      logger.err('Could not find androidPackage in pubspec.yaml.');
-      exit(ExitCode.config.code);
     }
   }
 
   @override
   Future<File> buildPatchArtifact() async {
+    final File aabFile;
     final flutterVersionString = await shorebirdFlutter.getVersionAndRevision();
     final buildProgress =
         logger.progress('Building patch with Flutter $flutterVersionString');
 
     try {
-      await artifactBuilder.buildAar(buildNumber: buildNumber);
+      aabFile =
+          await artifactBuilder.buildAppBundle(flavor: flavor, target: target);
       buildProgress.complete();
     } on ArtifactBuildException catch (error) {
-      buildProgress.fail('Failed to build: ${error.message}');
+      buildProgress.fail(error.message);
       exit(ExitCode.software.code);
     }
 
-    return File(
-      ShorebirdAndroidArtifacts.aarArtifactPath(
-        buildNumber: buildNumber,
-        packageName: shorebirdEnv.androidPackageName!,
-      ),
+    final patchArchsBuildDir = ArtifactManager.androidArchsDirectory(
+      projectRoot: projectRoot,
+      flavor: flavor,
     );
+
+    if (patchArchsBuildDir == null) {
+      logger
+        ..err('Cannot find patch build artifacts.')
+        ..info(
+          '''
+Please run `shorebird cache clean` and try again. If the issue persists, please
+file a bug report at https://github.com/shorebirdtech/shorebird/issues/new.
+
+Looked in:
+  - build/app/intermediates/stripped_native_libs/stripReleaseDebugSymbols/release/out/lib
+  - build/app/intermediates/stripped_native_libs/strip{flavor}ReleaseDebugSymbols/{flavor}Release/out/lib
+  - build/app/intermediates/stripped_native_libs/release/out/lib
+  - build/app/intermediates/stripped_native_libs/{flavor}Release/out/lib''',
+        );
+      exit(ExitCode.software.code);
+    }
+    return aabFile;
   }
 
   @override
@@ -96,7 +104,6 @@ class AarPatcher extends Patcher {
       architectures: AndroidArch.availableAndroidArchs,
       platform: releaseType.releasePlatform,
     );
-
     final releaseArtifactPaths = <Arch, String>{};
     final downloadReleaseArtifactProgress = logger.progress(
       'Downloading release artifacts',
@@ -116,31 +123,33 @@ class AarPatcher extends Patcher {
 
     downloadReleaseArtifactProgress.complete();
 
-    final extractedAarDirectory = await shorebirdAndroidArtifacts.extractAar(
-      packageName: shorebirdEnv.androidPackageName!,
-      buildNumber: buildNumber,
-      unzipFn: extractFileToDisk,
+    final patchArchsBuildDir = ArtifactManager.androidArchsDirectory(
+      projectRoot: projectRoot,
+      flavor: flavor,
     );
-    final patchArtifactBundles = <Arch, PatchArtifactBundle>{};
+    if (patchArchsBuildDir == null) {
+      logger.err('Could not find patch artifacts');
+      exit(ExitCode.software.code);
+    }
 
-    final createDiffProgress = logger.progress('Creating artifacts');
+    final patchArtifactBundles = <Arch, PatchArtifactBundle>{};
+    final createDiffProgress = logger.progress('Creating patch artifacts');
     for (final releaseArtifactPath in releaseArtifactPaths.entries) {
       final arch = releaseArtifactPath.key;
-      final artifactPath = p.join(
-        extractedAarDirectory.path,
-        'jni',
+      final patchArtifactPath = p.join(
+        patchArchsBuildDir.path,
         arch.androidBuildPath,
         'libapp.so',
       );
-      logger.detail('Creating artifact for $artifactPath');
-      final patchArtifact = File(artifactPath);
+      logger.detail('Creating artifact for $patchArtifactPath');
+      final patchArtifact = File(patchArtifactPath);
       final hash = sha256.convert(await patchArtifact.readAsBytes()).toString();
       try {
         final diffPath = await artifactManager.createDiff(
           releaseArtifactPath: releaseArtifactPath.value,
-          patchArtifactPath: artifactPath,
+          patchArtifactPath: patchArtifactPath,
         );
-        patchArtifactBundles[arch] = PatchArtifactBundle(
+        patchArtifactBundles[releaseArtifactPath.key] = PatchArtifactBundle(
           arch: arch.arch,
           path: diffPath,
           hash: hash,
@@ -152,15 +161,13 @@ class AarPatcher extends Patcher {
       }
     }
     createDiffProgress.complete();
-
     return patchArtifactBundles;
   }
 
   @override
-  Future<String> extractReleaseVersionFromArtifact(File artifact) {
-    // Not implemented - release verison must be specified by the user.
-    throw UnimplementedError(
-      'Release version must be specified using --release-version.',
+  Future<String> extractReleaseVersionFromArtifact(File artifact) async {
+    return shorebirdAndroidArtifacts.extractReleaseVersionFromAppBundle(
+      artifact.path,
     );
   }
