+++ packages/shorebird_cli/lib/src/commands/patch/patch.dart
@@ -1,5 +1,6 @@
-export 'patch_aar_command.dart';
-export 'patch_android_command.dart';
+export 'aar_patcher.dart';
+export 'android_patcher.dart';
+export 'ios_framework_patcher.dart';
+export 'ios_patcher.dart';
 export 'patch_command.dart';
-export 'patch_ios_command.dart';
-export 'patch_ios_framework_command.dart';
+export 'patcher.dart';
