--- packages/shorebird_cli/lib/src/commands/patch_new/patch_new.dart
@@ -1,6 +1,0 @@
-export 'aar_patcher.dart';
-export 'android_patcher.dart';
-export 'ios_framework_patcher.dart';
-export 'ios_patcher.dart';
-export 'patch_new_command.dart';
-export 'patcher.dart';
