+++ packages/shorebird_cli/test/src/commands/patch/aar_patcher_test.dart
@@ -8,20 +8,18 @@ import 'package:shorebird_cli/src/archive_analysis/archive_analysis.dart';
 import 'package:shorebird_cli/src/artifact_builder.dart';
 import 'package:shorebird_cli/src/artifact_manager.dart';
 import 'package:shorebird_cli/src/code_push_client_wrapper.dart';
-import 'package:shorebird_cli/src/commands/patch_new/patch_new.dart';
-import 'package:shorebird_cli/src/doctor.dart';
+import 'package:shorebird_cli/src/commands/patch/patch.dart';
 import 'package:shorebird_cli/src/engine_config.dart';
 import 'package:shorebird_cli/src/logger.dart';
 import 'package:shorebird_cli/src/patch_diff_checker.dart';
 import 'package:shorebird_cli/src/platform.dart';
 import 'package:shorebird_cli/src/platform/platform.dart';
+import 'package:shorebird_cli/src/release_type.dart';
 import 'package:shorebird_cli/src/shorebird_android_artifacts.dart';
 import 'package:shorebird_cli/src/shorebird_env.dart';
 import 'package:shorebird_cli/src/shorebird_flutter.dart';
-import 'package:shorebird_cli/src/shorebird_process.dart';
 import 'package:shorebird_cli/src/shorebird_validator.dart';
 import 'package:shorebird_cli/src/third_party/flutter_tools/lib/flutter_tools.dart';
-import 'package:shorebird_cli/src/validators/validators.dart';
 import 'package:shorebird_cli/src/version.dart';
 import 'package:shorebird_code_push_client/shorebird_code_push_client.dart';
 import 'package:test/test.dart';
@@ -30,42 +28,24 @@ import '../../matchers.dart';
 import '../../mocks.dart';
 
 void main() {
-  group(AndroidPatcher, () {
+  group(AarPatcher, () {
+    const packageName = 'com.example.my_flutter_module';
+    const buildNumber = '1.0';
+
     late ArgResults argResults;
     late ArtifactBuilder artifactBuilder;
     late ArtifactManager artifactManager;
     late CodePushClientWrapper codePushClientWrapper;
-    late Doctor doctor;
-    late Platform platform;
     late Directory projectRoot;
     late ShorebirdLogger logger;
+    late Platform platform;
     late Progress progress;
-    late ShorebirdFlutterValidator flutterValidator;
-    late ShorebirdProcess shorebirdProcess;
     late ShorebirdEnv shorebirdEnv;
     late ShorebirdFlutter shorebirdFlutter;
     late ShorebirdValidator shorebirdValidator;
     late ShorebirdAndroidArtifacts shorebirdAndroidArtifacts;
 
-    late AndroidPatcher patcher;
-
-    void setUpProjectRootArtifacts({String? flavor}) {
-      for (final archMetadata in Arch.values) {
-        final artifactPath = p.join(
-          projectRoot.path,
-          'build',
-          'app',
-          'intermediates',
-          'stripped_native_libs',
-          flavor != null ? '${flavor}Release' : 'release',
-          'out',
-          'lib',
-          archMetadata.androidBuildPath,
-          'libapp.so',
-        );
-        File(artifactPath).createSync(recursive: true);
-      }
-    }
+    late AarPatcher patcher;
 
     R runWithOverrides<R>(R Function() body) {
       return runScoped(
@@ -74,18 +54,29 @@ void main() {
           artifactBuilderRef.overrideWith(() => artifactBuilder),
           artifactManagerRef.overrideWith(() => artifactManager),
           codePushClientWrapperRef.overrideWith(() => codePushClientWrapper),
-          doctorRef.overrideWith(() => doctor),
           engineConfigRef.overrideWith(() => const EngineConfig.empty()),
           loggerRef.overrideWith(() => logger),
           platformRef.overrideWith(() => platform),
-          processRef.overrideWith(() => shorebirdProcess),
           shorebirdEnvRef.overrideWith(() => shorebirdEnv),
           shorebirdFlutterRef.overrideWith(() => shorebirdFlutter),
           shorebirdValidatorRef.overrideWith(() => shorebirdValidator),
-          shorebirdAndroidArtifactsRef
-              .overrideWith(() => shorebirdAndroidArtifacts),
+          shorebirdAndroidArtifactsRef.overrideWith(
+            () => shorebirdAndroidArtifacts,
+          ),
         },
       );
+    }
+
+    void setUpExtractedAarDirectory(Directory root) {
+      for (final archMetadata in Arch.values) {
+        final artifactPath = p.join(
+          root.path,
+          'jni',
+          archMetadata.androidBuildPath,
+          'libapp.so',
+        );
+        File(artifactPath).createSync(recursive: true);
+      }
     }
 
     setUpAll(() {
@@ -102,29 +93,35 @@ void main() {
       artifactBuilder = MockArtifactBuilder();
       artifactManager = MockArtifactManager();
       codePushClientWrapper = MockCodePushClientWrapper();
-      doctor = MockDoctor();
       platform = MockPlatform();
       progress = MockProgress();
       projectRoot = Directory.systemTemp.createTempSync();
       logger = MockShorebirdLogger();
-      flutterValidator = MockShorebirdFlutterValidator();
-      shorebirdProcess = MockShorebirdProcess();
       shorebirdEnv = MockShorebirdEnv();
       shorebirdFlutter = MockShorebirdFlutter();
       shorebirdValidator = MockShorebirdValidator();
       shorebirdAndroidArtifacts = MockShorebirdAndroidArtifacts();
 
+      when(() => argResults['build-number']).thenReturn('1.0');
+
       when(() => logger.progress(any())).thenReturn(progress);
 
+      when(() => shorebirdEnv.androidPackageName).thenReturn(packageName);
       when(
         () => shorebirdEnv.getShorebirdProjectRoot(),
       ).thenReturn(projectRoot);
 
-      patcher = AndroidPatcher(
-        argResults: argResults,
-        flavor: null,
-        target: null,
-      );
+      patcher = AarPatcher(argResults: argResults, flavor: null, target: null);
+    });
+
+    group('buildNumber', () {
+      setUp(() {
+        when(() => argResults['build-number']).thenReturn(buildNumber);
+      });
+
+      test('is the value of the build-number argument', () {
+        expect(patcher.buildNumber, buildNumber);
+      });
     });
 
     group('archiveDiffer', () {
@@ -133,25 +130,19 @@ void main() {
       });
     });
 
-    group('primaryReleaseArtifactArch', () {
-      test('is "aab"', () {
-        expect(patcher.primaryReleaseArtifactArch, equals('aab'));
+    group('releaseType', () {
+      test('is aar', () {
+        expect(patcher.releaseType, ReleaseType.aar);
       });
     });
 
-    group('assertArgsAreValid', () {
-      test('does nothing', () async {
-        await expectLater(patcher.assertArgsAreValid(), completes);
+    group('primaryReleaseArtifactArch', () {
+      test('is aar', () {
+        expect(patcher.primaryReleaseArtifactArch, equals('aar'));
       });
     });
 
     group('assertPreconditions', () {
-      setUp(() {
-        when(() => doctor.androidCommandValidators)
-            .thenReturn([flutterValidator]);
-        when(flutterValidator.validate).thenAnswer((_) async => []);
-      });
-
       group('when validation succeeds', () {
         setUp(() {
           when(
@@ -171,6 +162,23 @@ void main() {
             () => runWithOverrides(patcher.assertPreconditions),
             returnsNormally,
           );
+        });
+
+        group('when androidPackageName is null', () {
+          setUp(() {
+            when(() => shorebirdEnv.androidPackageName).thenReturn(null);
+          });
+
+          test('logs error and exits with code 64', () async {
+            await expectLater(
+              () => runWithOverrides(patcher.assertPreconditions),
+              exitsWithCode(ExitCode.config),
+            );
+            verify(
+              () =>
+                  logger.err('Could not find androidPackage in pubspec.yaml.'),
+            ).called(1);
+          });
         });
       });
 
@@ -205,7 +213,6 @@ void main() {
             () => shorebirdValidator.validatePreconditions(
               checkUserIsAuthenticated: true,
               checkShorebirdInitialized: true,
-              validators: [flutterValidator],
             ),
           ).called(1);
         });
@@ -214,20 +221,11 @@ void main() {
 
     group('buildPatchArtifact', () {
       const flutterVersionAndRevision = '3.10.6 (83305b5088)';
-      late File aabFile;
 
       setUp(() {
-        aabFile = File('');
         when(
           () => shorebirdFlutter.getVersionAndRevision(),
         ).thenAnswer((_) async => flutterVersionAndRevision);
-        when(
-          () => artifactBuilder.buildAppBundle(
-            flavor: any(named: 'flavor'),
-            target: any(named: 'target'),
-            targetPlatforms: any(named: 'targetPlatforms'),
-          ),
-        ).thenAnswer((_) async => aabFile);
       });
 
       group('when build fails', () {
@@ -235,12 +233,10 @@ void main() {
 
         setUp(() {
           when(
-            () => artifactBuilder.buildAppBundle(
-              flavor: any(named: 'flavor'),
-              target: any(named: 'target'),
+            () => artifactBuilder.buildAar(
+              buildNumber: any(named: 'buildNumber'),
             ),
           ).thenThrow(exception);
-          when(() => logger.progress(any())).thenReturn(progress);
         });
 
         test('logs error and exits with code 70', () async {
@@ -249,51 +245,54 @@ void main() {
             exitsWithCode(ExitCode.software),
           );
 
-          verify(() => progress.fail('error')).called(1);
-        });
-      });
-
-      group('when patch artifacts cannot be found', () {
-        test('logs error and exits with code 70', () async {
-          await expectLater(
-            () => runWithOverrides(patcher.buildPatchArtifact),
-            exitsWithCode(ExitCode.software),
-          );
-
-          verify(
-            () => logger.err('Cannot find patch build artifacts.'),
-          ).called(1);
-          verify(
-            () => logger.info(
-              '''
-Please run `shorebird cache clean` and try again. If the issue persists, please
-file a bug report at https://github.com/shorebirdtech/shorebird/issues/new.
-
-Looked in:
-  - build/app/intermediates/stripped_native_libs/stripReleaseDebugSymbols/release/out/lib
-  - build/app/intermediates/stripped_native_libs/strip{flavor}ReleaseDebugSymbols/{flavor}Release/out/lib
-  - build/app/intermediates/stripped_native_libs/release/out/lib
-  - build/app/intermediates/stripped_native_libs/{flavor}Release/out/lib''',
-            ),
-          ).called(1);
+          verify(() => progress.fail('Failed to build: error')).called(1);
         });
       });
 
       group('when build succeeds', () {
-        setUp(setUpProjectRootArtifacts);
+        setUp(() {
+          when(
+            () => artifactBuilder.buildAar(
+              buildNumber: any(named: 'buildNumber'),
+            ),
+          ).thenAnswer((_) async => {});
+        });
 
-        test('returns the aab file', () async {
-          final result = await runWithOverrides(patcher.buildPatchArtifact);
-          expect(result, equals(aabFile));
+        test('returns the aar artifact file', () async {
+          final artifact = await runWithOverrides(patcher.buildPatchArtifact);
+
+          expect(artifact, isA<File>());
+          expect(
+            artifact.path,
+            endsWith(
+              p.join(
+                'build',
+                'host',
+                'outputs',
+                'repo',
+                'com',
+                'example',
+                'my_flutter_module',
+                'flutter_release',
+                buildNumber,
+                'flutter_release-$buildNumber.aar',
+              ),
+            ),
+          );
+
+          verify(() => artifactBuilder.buildAar(buildNumber: buildNumber))
+              .called(1);
         });
       });
     });
 
     group('createPatchArtifacts', () {
+      const appId = 'appId';
       const arch = 'aarch64';
+      const releaseId = 1;
       const releaseArtifact = ReleaseArtifact(
         id: 0,
-        releaseId: 0,
+        releaseId: releaseId,
         arch: arch,
         platform: ReleasePlatform.android,
         hash: '#',
@@ -301,7 +300,15 @@ Looked in:
         url: 'https://example.com',
       );
 
+      late File releaseArtifactFile;
+      late Directory extractedAarDirectory;
+
       setUp(() {
+        releaseArtifactFile = File('');
+        when(() => artifactManager.downloadFile(any())).thenAnswer(
+          (_) async => releaseArtifactFile,
+        );
+
         when(
           () => codePushClientWrapper.getReleaseArtifacts(
             appId: any(named: 'appId'),
@@ -316,23 +323,42 @@ Looked in:
             Arch.x86_64: releaseArtifact,
           },
         );
-        when(() => artifactManager.downloadFile(any()))
-            .thenAnswer((_) async => File(''));
+
+        extractedAarDirectory = Directory.systemTemp.createTempSync();
+        setUpExtractedAarDirectory(extractedAarDirectory);
+        when(
+          () => shorebirdAndroidArtifacts.extractAar(
+            packageName: any(named: 'packageName'),
+            buildNumber: any(named: 'buildNumber'),
+            unzipFn: any(named: 'unzipFn'),
+          ),
+        ).thenAnswer((_) async => extractedAarDirectory);
+
+        when(
+          () => artifactManager.createDiff(
+            patchArtifactPath: any(named: 'patchArtifactPath'),
+            releaseArtifactPath: any(named: 'releaseArtifactPath'),
+          ),
+        ).thenAnswer((_) async {
+          final diffPath =
+              File(p.join(Directory.systemTemp.createTempSync().path, 'diff'))
+                ..createSync();
+          return diffPath.path;
+        });
       });
 
-      group('when release artifact fails to download', () {
+      group('when release artifact download fails', () {
+        final exception = Exception('error');
         setUp(() {
-          when(
-            () => artifactManager.downloadFile(any()),
-          ).thenThrow(Exception('error'));
+          when(() => artifactManager.downloadFile(any())).thenThrow(exception);
         });
 
         test('logs error and exits with code 70', () async {
           await expectLater(
             () => runWithOverrides(
               () => patcher.createPatchArtifacts(
-                appId: 'appId',
-                releaseId: 0,
+                appId: appId,
+                releaseId: releaseId,
               ),
             ),
             exitsWithCode(ExitCode.software),
@@ -342,26 +368,8 @@ Looked in:
         });
       });
 
-      group('when unable to find patch build artifacts', () {
-        test('logs error and exits with code 70', () async {
-          await expectLater(
-            () => runWithOverrides(
-              () => patcher.createPatchArtifacts(
-                appId: 'appId',
-                releaseId: 0,
-              ),
-            ),
-            exitsWithCode(ExitCode.software),
-          );
-
-          verify(() => logger.err('Could not find patch artifacts')).called(1);
-        });
-      });
-
-      group('when unable to create diffs', () {
+      group('when diff creation fails', () {
         setUp(() {
-          setUpProjectRootArtifacts();
-
           when(
             () => artifactManager.createDiff(
               releaseArtifactPath: any(named: 'releaseArtifactPath'),
@@ -374,8 +382,8 @@ Looked in:
           await expectLater(
             () => runWithOverrides(
               () => patcher.createPatchArtifacts(
-                appId: 'appId',
-                releaseId: 0,
+                appId: appId,
+                releaseId: releaseId,
               ),
             ),
             exitsWithCode(ExitCode.software),
@@ -385,68 +393,34 @@ Looked in:
         });
       });
 
-      group('when patch artifacts successfully created', () {
-        setUp(() {
-          setUpProjectRootArtifacts();
-          when(
-            () => artifactManager.createDiff(
-              releaseArtifactPath: any(named: 'releaseArtifactPath'),
-              patchArtifactPath: any(named: 'patchArtifactPath'),
-            ),
-          ).thenAnswer((_) async {
-            final tempDir = Directory.systemTemp.createTempSync();
-            final diffPath = p.join(tempDir.path, 'diff');
-            File(diffPath)
-              ..createSync()
-              ..writeAsStringSync('test');
-            return diffPath;
-          });
-        });
-
-        test('returns patch artifact bundles', () async {
-          final result = await runWithOverrides(
+      group('when successful', () {
+        test('returns map of archs to patch artifacts', () async {
+          final patchArtifacts = await runWithOverrides(
             () => patcher.createPatchArtifacts(
-              appId: 'appId',
-              releaseId: 0,
+              appId: appId,
+              releaseId: releaseId,
             ),
           );
 
-          expect(result, hasLength(Arch.values.length));
+          expect(patchArtifacts, hasLength(3));
+          expect(
+            patchArtifacts.keys,
+            containsAll([Arch.arm32, Arch.arm64, Arch.x86_64]),
+          );
+          expect(
+            patchArtifacts.values,
+            everyElement(isA<PatchArtifactBundle>()),
+          );
         });
       });
     });
 
     group('extractReleaseVersionFromArtifact', () {
-      setUp(() {
-        when(
-          () => shorebirdAndroidArtifacts.extractReleaseVersionFromAppBundle(
-            any(),
-          ),
-        ).thenAnswer((_) async => '1.0.0');
-      });
-
-      test(
-          '''returns value of shorebirdAndroidArtifacts.extractReleaseVersionFromAppBundle''',
-          () async {
+      test('throws UnimplementedError', () {
         expect(
-          await runWithOverrides(
-            () => patcher.extractReleaseVersionFromArtifact(File('')),
-          ),
-          equals('1.0.0'),
+          () => patcher.extractReleaseVersionFromArtifact(File('')),
+          throwsUnimplementedError,
         );
-      });
-    });
-
-    group('patchArtifactForDiffCheck', () {
-      late File aabFile;
-      setUp(() {
-        aabFile = File('');
-        when(
-          () => shorebirdAndroidArtifacts.findAab(
-            project: any(named: 'project'),
-            flavor: any(named: 'flavor'),
-          ),
-        ).thenReturn(aabFile);
       });
     });
