+++ packages/shorebird_cli/test/src/commands/patch/android_patcher_test.dart
@@ -8,18 +8,20 @@ import 'package:shorebird_cli/src/archive_analysis/archive_analysis.dart';
 import 'package:shorebird_cli/src/artifact_builder.dart';
 import 'package:shorebird_cli/src/artifact_manager.dart';
 import 'package:shorebird_cli/src/code_push_client_wrapper.dart';
-import 'package:shorebird_cli/src/commands/patch_new/patch_new.dart';
+import 'package:shorebird_cli/src/commands/patch/patch.dart';
+import 'package:shorebird_cli/src/doctor.dart';
 import 'package:shorebird_cli/src/engine_config.dart';
 import 'package:shorebird_cli/src/logger.dart';
 import 'package:shorebird_cli/src/patch_diff_checker.dart';
 import 'package:shorebird_cli/src/platform.dart';
 import 'package:shorebird_cli/src/platform/platform.dart';
-import 'package:shorebird_cli/src/release_type.dart';
 import 'package:shorebird_cli/src/shorebird_android_artifacts.dart';
 import 'package:shorebird_cli/src/shorebird_env.dart';
 import 'package:shorebird_cli/src/shorebird_flutter.dart';
+import 'package:shorebird_cli/src/shorebird_process.dart';
 import 'package:shorebird_cli/src/shorebird_validator.dart';
 import 'package:shorebird_cli/src/third_party/flutter_tools/lib/flutter_tools.dart';
+import 'package:shorebird_cli/src/validators/validators.dart';
 import 'package:shorebird_cli/src/version.dart';
 import 'package:shorebird_code_push_client/shorebird_code_push_client.dart';
 import 'package:test/test.dart';
@@ -28,24 +30,42 @@ import '../../matchers.dart';
 import '../../mocks.dart';
 
 void main() {
-  group(AarPatcher, () {
-    const packageName = 'com.example.my_flutter_module';
-    const buildNumber = '1.0';
-
+  group(AndroidPatcher, () {
     late ArgResults argResults;
     late ArtifactBuilder artifactBuilder;
     late ArtifactManager artifactManager;
     late CodePushClientWrapper codePushClientWrapper;
+    late Doctor doctor;
+    late Platform platform;
     late Directory projectRoot;
     late ShorebirdLogger logger;
-    late Platform platform;
     late Progress progress;
+    late ShorebirdFlutterValidator flutterValidator;
+    late ShorebirdProcess shorebirdProcess;
     late ShorebirdEnv shorebirdEnv;
     late ShorebirdFlutter shorebirdFlutter;
     late ShorebirdValidator shorebirdValidator;
     late ShorebirdAndroidArtifacts shorebirdAndroidArtifacts;
 
-    late AarPatcher patcher;
+    late AndroidPatcher patcher;
+
+    void setUpProjectRootArtifacts({String? flavor}) {
+      for (final archMetadata in Arch.values) {
+        final artifactPath = p.join(
+          projectRoot.path,
+          'build',
+          'app',
+          'intermediates',
+          'stripped_native_libs',
+          flavor != null ? '${flavor}Release' : 'release',
+          'out',
+          'lib',
+          archMetadata.androidBuildPath,
+          'libapp.so',
+        );
+        File(artifactPath).createSync(recursive: true);
+      }
+    }
 
     R runWithOverrides<R>(R Function() body) {
       return runScoped(
@@ -54,29 +74,18 @@ void main() {
           artifactBuilderRef.overrideWith(() => artifactBuilder),
           artifactManagerRef.overrideWith(() => artifactManager),
           codePushClientWrapperRef.overrideWith(() => codePushClientWrapper),
+          doctorRef.overrideWith(() => doctor),
           engineConfigRef.overrideWith(() => const EngineConfig.empty()),
           loggerRef.overrideWith(() => logger),
           platformRef.overrideWith(() => platform),
+          processRef.overrideWith(() => shorebirdProcess),
           shorebirdEnvRef.overrideWith(() => shorebirdEnv),
           shorebirdFlutterRef.overrideWith(() => shorebirdFlutter),
           shorebirdValidatorRef.overrideWith(() => shorebirdValidator),
-          shorebirdAndroidArtifactsRef.overrideWith(
-            () => shorebirdAndroidArtifacts,
-          ),
+          shorebirdAndroidArtifactsRef
+              .overrideWith(() => shorebirdAndroidArtifacts),
         },
       );
-    }
-
-    void setUpExtractedAarDirectory(Directory root) {
-      for (final archMetadata in Arch.values) {
-        final artifactPath = p.join(
-          root.path,
-          'jni',
-          archMetadata.androidBuildPath,
-          'libapp.so',
-        );
-        File(artifactPath).createSync(recursive: true);
-      }
     }
 
     setUpAll(() {
@@ -93,35 +102,29 @@ void main() {
       artifactBuilder = MockArtifactBuilder();
       artifactManager = MockArtifactManager();
       codePushClientWrapper = MockCodePushClientWrapper();
+      doctor = MockDoctor();
       platform = MockPlatform();
       progress = MockProgress();
       projectRoot = Directory.systemTemp.createTempSync();
       logger = MockShorebirdLogger();
+      flutterValidator = MockShorebirdFlutterValidator();
+      shorebirdProcess = MockShorebirdProcess();
       shorebirdEnv = MockShorebirdEnv();
       shorebirdFlutter = MockShorebirdFlutter();
       shorebirdValidator = MockShorebirdValidator();
       shorebirdAndroidArtifacts = MockShorebirdAndroidArtifacts();
 
-      when(() => argResults['build-number']).thenReturn('1.0');
-
       when(() => logger.progress(any())).thenReturn(progress);
 
-      when(() => shorebirdEnv.androidPackageName).thenReturn(packageName);
       when(
         () => shorebirdEnv.getShorebirdProjectRoot(),
       ).thenReturn(projectRoot);
 
-      patcher = AarPatcher(argResults: argResults, flavor: null, target: null);
-    });
-
-    group('buildNumber', () {
-      setUp(() {
-        when(() => argResults['build-number']).thenReturn(buildNumber);
-      });
-
-      test('is the value of the build-number argument', () {
-        expect(patcher.buildNumber, buildNumber);
-      });
+      patcher = AndroidPatcher(
+        argResults: argResults,
+        flavor: null,
+        target: null,
+      );
     });
 
     group('archiveDiffer', () {
@@ -130,19 +133,25 @@ void main() {
       });
     });
 
-    group('releaseType', () {
-      test('is aar', () {
-        expect(patcher.releaseType, ReleaseType.aar);
+    group('primaryReleaseArtifactArch', () {
+      test('is "aab"', () {
+        expect(patcher.primaryReleaseArtifactArch, equals('aab'));
       });
     });
 
-    group('primaryReleaseArtifactArch', () {
-      test('is aar', () {
-        expect(patcher.primaryReleaseArtifactArch, equals('aar'));
+    group('assertArgsAreValid', () {
+      test('does nothing', () async {
+        await expectLater(patcher.assertArgsAreValid(), completes);
       });
     });
 
     group('assertPreconditions', () {
+      setUp(() {
+        when(() => doctor.androidCommandValidators)
+            .thenReturn([flutterValidator]);
+        when(flutterValidator.validate).thenAnswer((_) async => []);
+      });
+
       group('when validation succeeds', () {
         setUp(() {
           when(
@@ -162,23 +171,6 @@ void main() {
             () => runWithOverrides(patcher.assertPreconditions),
             returnsNormally,
           );
-        });
-
-        group('when androidPackageName is null', () {
-          setUp(() {
-            when(() => shorebirdEnv.androidPackageName).thenReturn(null);
-          });
-
-          test('logs error and exits with code 64', () async {
-            await expectLater(
-              () => runWithOverrides(patcher.assertPreconditions),
-              exitsWithCode(ExitCode.config),
-            );
-            verify(
-              () =>
-                  logger.err('Could not find androidPackage in pubspec.yaml.'),
-            ).called(1);
-          });
         });
       });
 
@@ -213,6 +205,7 @@ void main() {
             () => shorebirdValidator.validatePreconditions(
               checkUserIsAuthenticated: true,
               checkShorebirdInitialized: true,
+              validators: [flutterValidator],
             ),
           ).called(1);
         });
@@ -221,11 +214,20 @@ void main() {
 
     group('buildPatchArtifact', () {
       const flutterVersionAndRevision = '3.10.6 (83305b5088)';
+      late File aabFile;
 
       setUp(() {
+        aabFile = File('');
         when(
           () => shorebirdFlutter.getVersionAndRevision(),
         ).thenAnswer((_) async => flutterVersionAndRevision);
+        when(
+          () => artifactBuilder.buildAppBundle(
+            flavor: any(named: 'flavor'),
+            target: any(named: 'target'),
+            targetPlatforms: any(named: 'targetPlatforms'),
+          ),
+        ).thenAnswer((_) async => aabFile);
       });
 
       group('when build fails', () {
@@ -233,10 +235,12 @@ void main() {
 
         setUp(() {
           when(
-            () => artifactBuilder.buildAar(
-              buildNumber: any(named: 'buildNumber'),
+            () => artifactBuilder.buildAppBundle(
+              flavor: any(named: 'flavor'),
+              target: any(named: 'target'),
             ),
           ).thenThrow(exception);
+          when(() => logger.progress(any())).thenReturn(progress);
         });
 
         test('logs error and exits with code 70', () async {
@@ -245,54 +249,51 @@ void main() {
             exitsWithCode(ExitCode.software),
           );
 
-          verify(() => progress.fail('Failed to build: error')).called(1);
+          verify(() => progress.fail('error')).called(1);
+        });
+      });
+
+      group('when patch artifacts cannot be found', () {
+        test('logs error and exits with code 70', () async {
+          await expectLater(
+            () => runWithOverrides(patcher.buildPatchArtifact),
+            exitsWithCode(ExitCode.software),
+          );
+
+          verify(
+            () => logger.err('Cannot find patch build artifacts.'),
+          ).called(1);
+          verify(
+            () => logger.info(
+              '''
+Please run `shorebird cache clean` and try again. If the issue persists, please
+file a bug report at https://github.com/shorebirdtech/shorebird/issues/new.
+
+Looked in:
+  - build/app/intermediates/stripped_native_libs/stripReleaseDebugSymbols/release/out/lib
+  - build/app/intermediates/stripped_native_libs/strip{flavor}ReleaseDebugSymbols/{flavor}Release/out/lib
+  - build/app/intermediates/stripped_native_libs/release/out/lib
+  - build/app/intermediates/stripped_native_libs/{flavor}Release/out/lib''',
+            ),
+          ).called(1);
         });
       });
 
       group('when build succeeds', () {
-        setUp(() {
-          when(
-            () => artifactBuilder.buildAar(
-              buildNumber: any(named: 'buildNumber'),
-            ),
-          ).thenAnswer((_) async => {});
-        });
+        setUp(setUpProjectRootArtifacts);
 
-        test('returns the aar artifact file', () async {
-          final artifact = await runWithOverrides(patcher.buildPatchArtifact);
-
-          expect(artifact, isA<File>());
-          expect(
-            artifact.path,
-            endsWith(
-              p.join(
-                'build',
-                'host',
-                'outputs',
-                'repo',
-                'com',
-                'example',
-                'my_flutter_module',
-                'flutter_release',
-                buildNumber,
-                'flutter_release-$buildNumber.aar',
-              ),
-            ),
-          );
-
-          verify(() => artifactBuilder.buildAar(buildNumber: buildNumber))
-              .called(1);
+        test('returns the aab file', () async {
+          final result = await runWithOverrides(patcher.buildPatchArtifact);
+          expect(result, equals(aabFile));
         });
       });
     });
 
     group('createPatchArtifacts', () {
-      const appId = 'appId';
       const arch = 'aarch64';
-      const releaseId = 1;
       const releaseArtifact = ReleaseArtifact(
         id: 0,
-        releaseId: releaseId,
+        releaseId: 0,
         arch: arch,
         platform: ReleasePlatform.android,
         hash: '#',
@@ -300,15 +301,7 @@ void main() {
         url: 'https://example.com',
       );
 
-      late File releaseArtifactFile;
-      late Directory extractedAarDirectory;
-
       setUp(() {
-        releaseArtifactFile = File('');
-        when(() => artifactManager.downloadFile(any())).thenAnswer(
-          (_) async => releaseArtifactFile,
-        );
-
         when(
           () => codePushClientWrapper.getReleaseArtifacts(
             appId: any(named: 'appId'),
@@ -323,42 +316,23 @@ void main() {
             Arch.x86_64: releaseArtifact,
           },
         );
-
-        extractedAarDirectory = Directory.systemTemp.createTempSync();
-        setUpExtractedAarDirectory(extractedAarDirectory);
-        when(
-          () => shorebirdAndroidArtifacts.extractAar(
-            packageName: any(named: 'packageName'),
-            buildNumber: any(named: 'buildNumber'),
-            unzipFn: any(named: 'unzipFn'),
-          ),
-        ).thenAnswer((_) async => extractedAarDirectory);
-
-        when(
-          () => artifactManager.createDiff(
-            patchArtifactPath: any(named: 'patchArtifactPath'),
-            releaseArtifactPath: any(named: 'releaseArtifactPath'),
-          ),
-        ).thenAnswer((_) async {
-          final diffPath =
-              File(p.join(Directory.systemTemp.createTempSync().path, 'diff'))
-                ..createSync();
-          return diffPath.path;
-        });
+        when(() => artifactManager.downloadFile(any()))
+            .thenAnswer((_) async => File(''));
       });
 
-      group('when release artifact download fails', () {
-        final exception = Exception('error');
+      group('when release artifact fails to download', () {
         setUp(() {
-          when(() => artifactManager.downloadFile(any())).thenThrow(exception);
+          when(
+            () => artifactManager.downloadFile(any()),
+          ).thenThrow(Exception('error'));
         });
 
         test('logs error and exits with code 70', () async {
           await expectLater(
             () => runWithOverrides(
               () => patcher.createPatchArtifacts(
-                appId: appId,
-                releaseId: releaseId,
+                appId: 'appId',
+                releaseId: 0,
               ),
             ),
             exitsWithCode(ExitCode.software),
@@ -368,8 +342,26 @@ void main() {
         });
       });
 
-      group('when diff creation fails', () {
+      group('when unable to find patch build artifacts', () {
+        test('logs error and exits with code 70', () async {
+          await expectLater(
+            () => runWithOverrides(
+              () => patcher.createPatchArtifacts(
+                appId: 'appId',
+                releaseId: 0,
+              ),
+            ),
+            exitsWithCode(ExitCode.software),
+          );
+
+          verify(() => logger.err('Could not find patch artifacts')).called(1);
+        });
+      });
+
+      group('when unable to create diffs', () {
         setUp(() {
+          setUpProjectRootArtifacts();
+
           when(
             () => artifactManager.createDiff(
               releaseArtifactPath: any(named: 'releaseArtifactPath'),
@@ -382,8 +374,8 @@ void main() {
           await expectLater(
             () => runWithOverrides(
               () => patcher.createPatchArtifacts(
-                appId: appId,
-                releaseId: releaseId,
+                appId: 'appId',
+                releaseId: 0,
               ),
             ),
             exitsWithCode(ExitCode.software),
@@ -393,34 +385,68 @@ void main() {
         });
       });
 
-      group('when successful', () {
-        test('returns map of archs to patch artifacts', () async {
-          final patchArtifacts = await runWithOverrides(
+      group('when patch artifacts successfully created', () {
+        setUp(() {
+          setUpProjectRootArtifacts();
+          when(
+            () => artifactManager.createDiff(
+              releaseArtifactPath: any(named: 'releaseArtifactPath'),
+              patchArtifactPath: any(named: 'patchArtifactPath'),
+            ),
+          ).thenAnswer((_) async {
+            final tempDir = Directory.systemTemp.createTempSync();
+            final diffPath = p.join(tempDir.path, 'diff');
+            File(diffPath)
+              ..createSync()
+              ..writeAsStringSync('test');
+            return diffPath;
+          });
+        });
+
+        test('returns patch artifact bundles', () async {
+          final result = await runWithOverrides(
             () => patcher.createPatchArtifacts(
-              appId: appId,
-              releaseId: releaseId,
+              appId: 'appId',
+              releaseId: 0,
             ),
           );
 
-          expect(patchArtifacts, hasLength(3));
-          expect(
-            patchArtifacts.keys,
-            containsAll([Arch.arm32, Arch.arm64, Arch.x86_64]),
-          );
-          expect(
-            patchArtifacts.values,
-            everyElement(isA<PatchArtifactBundle>()),
-          );
+          expect(result, hasLength(Arch.values.length));
         });
       });
     });
 
     group('extractReleaseVersionFromArtifact', () {
-      test('throws UnimplementedError', () {
+      setUp(() {
+        when(
+          () => shorebirdAndroidArtifacts.extractReleaseVersionFromAppBundle(
+            any(),
+          ),
+        ).thenAnswer((_) async => '1.0.0');
+      });
+
+      test(
+          '''returns value of shorebirdAndroidArtifacts.extractReleaseVersionFromAppBundle''',
+          () async {
         expect(
-          () => patcher.extractReleaseVersionFromArtifact(File('')),
-          throwsUnimplementedError,
+          await runWithOverrides(
+            () => patcher.extractReleaseVersionFromArtifact(File('')),
+          ),
+          equals('1.0.0'),
         );
+      });
+    });
+
+    group('patchArtifactForDiffCheck', () {
+      late File aabFile;
+      setUp(() {
+        aabFile = File('');
+        when(
+          () => shorebirdAndroidArtifacts.findAab(
+            project: any(named: 'project'),
+            flavor: any(named: 'flavor'),
+          ),
+        ).thenReturn(aabFile);
       });
     });
